+++ packages/core/src/utilities/loadImageToCanvas.ts
@@ -65,6 +65,7 @@ export default function loadImageToCanvas(
       targetBuffer: {
         type: useNorm16Texture ? undefined : 'Float32Array',
       },
+      useRGBA: true,
       preScale: {
         enabled: true,
       },