+++ packages/dicomImageLoader/src/types/ImageFrame.ts
@@ -1,3 +1,5 @@
+import PixelDataTypedArray from './PixelDataTypedArray';
+
 interface ImageFrame {
   samplesPerPixel: number;
   photometricInterpretation: string;
@@ -16,17 +18,22 @@ interface ImageFrame {
   greenPaletteColorLookupTableData: number[];
   bluePaletteColorLookupTableData: number[];
   // populated later after decoding
-  pixelData:
-    | Float32Array
-    | Int16Array
-    | Uint16Array
-    | Uint8Array
-    | Uint8ClampedArray
-    | undefined;
+  pixelData: PixelDataTypedArray;
   imageData?: ImageData;
   decodeTimeInMS?: number;
   pixelDataLength?: number;
-  preScale?: any;
+  preScale?: {
+    enabled?: boolean;
+    scalingParameters?: {
+      intercept: number;
+      slope: number;
+      modality?: string;
+      suvbw?: number;
+    };
+    scaled?: boolean;
+  };
+  minAfterScale?: number;
+  maxAfterScale?: number;
 }
 
 export default ImageFrame;