+++ packages/streaming-image-volume-loader/src/helpers/scaleArray.ts
@@ -14,11 +14,7 @@ export default function scaleArray(
   const arrayLength = array.length;
   const { rescaleSlope, rescaleIntercept, suvbw } = scalingParameters;
 
-  if (scalingParameters.modality === 'PT') {
-    if (typeof suvbw !== 'number') {
-      return array;
-    }
-
+  if (scalingParameters.modality === 'PT' && typeof suvbw === 'number') {
     for (let i = 0; i < arrayLength; i++) {
       array[i] = suvbw * (array[i] * rescaleSlope + rescaleIntercept);
     }