+++ packages/tools/src/eventListeners/mouse/mouseDownListener.ts
@@ -130,9 +130,11 @@ function mouseDownListener(evt: MouseEvent) {
     return;
   }
 
+  // Only left single button click can be doubled up, so fire immediately
+  // on anything except left double click.
   doubleClickState.doubleClickTimeout = setTimeout(
     _doStateMouseDownAndUp,
-    DOUBLE_CLICK_TOLERANCE_MS
+    evt.buttons === 1 ? DOUBLE_CLICK_TOLERANCE_MS : 0
   );
 
   // First mouse down of a potential double click. So save it and start
@@ -142,7 +144,7 @@ function mouseDownListener(evt: MouseEvent) {
 
   state.element = <HTMLDivElement>evt.currentTarget;
 
-  state.mouseButton = evt.button;
+  state.mouseButton = evt.buttons;
 
   const enabledElement = getEnabledElement(state.element);
   const { renderingEngineId, viewportId } = enabledElement;